+++ src/PixelDissimilarity.hpp
@@ -5,7 +5,7 @@
 template <class Pixel> class PixelDissimilarity {
   public:
     // Typedef for a pointer to a dissimilarity function
-    typedef double (PixelDissimilarity<Pixel>::*DissimilarityFunc)(ImgIdx index1, ImgIdx index2) const;
+    typedef float (PixelDissimilarity<Pixel>::*DissimilarityFunc)(ImgIdx index1, ImgIdx index2) const;
 
     // // Constructor that accepts a function pointer
     PixelDissimilarity() = default;
@@ -13,17 +13,17 @@ template <class Pixel> class PixelDissimilarity {
         : _image(image_), _imageSize(imageSize_), _channels(channels_), dissimilarityFunc(func) {}
 
     // Function to compute dissimilarity using the selected method
-    double computeDissimilarity(ImgIdx index1, ImgIdx index2) const {
+    float computeDissimilarity(ImgIdx index1, ImgIdx index2) const {
         return (this->*dissimilarityFunc)(index1, index2);
     }
 
-    double maximumDissmilarity() const {
-        return (double)std::numeric_limits<Pixel>::max() - (double)std::numeric_limits<Pixel>::min();
+    float maximumDissmilarity() const {
+        return (float)std::numeric_limits<Pixel>::max() - (float)std::numeric_limits<Pixel>::min();
     }
 
-    double L1(ImgIdx index1, ImgIdx index2) const;
-    double L2(ImgIdx index1, ImgIdx index2) const;
-    double LInfinity(ImgIdx index1, ImgIdx index2) const;
+    float L1(ImgIdx index1, ImgIdx index2) const;
+    float L2(ImgIdx index1, ImgIdx index2) const;
+    float LInfinity(ImgIdx index1, ImgIdx index2) const;
 
   private:
     const Pixel *_image = nullptr;
